pyproject.toml
[project]
name = "kernels"
version = "0.1.0"
description = "GPU kernels and benchmarking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "modal>=1.3.4",
    "cuda-tile; sys_platform == 'linux'",
    "nvidia-cutlass-dsl[cu13]; sys_platform == 'linux'",
    "quack-kernels; sys_platform == 'linux'",
    "apache-tvm-ffi; sys_platform == 'linux'",
    "torch",
    "triton; sys_platform == 'linux'",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu130", marker = "sys_platform == 'linux'" },
    { index = "pypi", marker = "sys_platform != 'linux'" },
]
